refactor(nat/gateway): move the data source to the new dir
