+++ huaweicloud/provider.go
@@ -481,7 +481,8 @@ func Provider() *schema.Provider {
 			"huaweicloud_elb_flavors":     elb.DataSourceElbFlavorsV3(),
 			"huaweicloud_elb_pools":       elb.DataSourcePools(),
 
-			"huaweicloud_nat_gateway":          DataSourceNatGatewayV2(),
+			"huaweicloud_nat_gateway": nat.DataSourceNatGatewayV2(),
+
 			"huaweicloud_networking_port":      vpc.DataSourceNetworkingPortV2(),
 			"huaweicloud_networking_secgroup":  DataSourceNetworkingSecGroup(),
 			"huaweicloud_networking_secgroups": vpc.DataSourceNetworkingSecGroups(),
